not working

+++ sii/includes/byte.h
@@ -1,9 +1,11 @@
 #ifndef __BYTE_H__
 #define __BYTE_H__
 
+#include <inttypes.h>
+
 typedef unsigned char BYTE;
 
-int getBit(BYTE byte, int whichBit);
-BYTE overrideLessSignificantBit(BYTE byte, int bit);
+int getBit(BYTE byte, uint8_t whichBit);
+BYTE overrideLessSignificantBit(BYTE byte, uint8_t bit);
 
 #endif
